serialize weapon condition

+++ Assets/Scripts/Data/EquipmentModel.cs
@@ -1,3 +1,3 @@
 ﻿public class EquipmentModel : ItemModel {
-    public int durability;
+    public float durability;
 }
